moved email content to a separate file
